third_party/blink/web_tests/virtual/popup-disabled/popup-disabled.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src='../../resources/testharness.js'></script>

<div id=elements>
  <popup id=popup_element>This content should be visible, if HTMLPopupElement is disabled</popup>
  <div popup=popup id=popup_attribute>This content should be visible, if HTMLPopupAttribute is disabled</div>
</div>

<script>
  if (window.testRunner)
    testRunner.dumpAsText();

  function popupVisible(popup) {
    return !!(popup.offsetWidth || popup.offsetHeight || popup.getClientRects().length);
  }

  const elements = document.getElementById('elements');
  const popupElement = elements.querySelector('popup');
  test(() => {
    assert_throws_dom("SyntaxError",() => {elements.matches(':popup-open')},"The :popup-open pseudo class shouldn't be available");
    assert_true(!!popupElement);
    assert_true(popupElement instanceof HTMLUnknownElement, '<popup> should be an instance of HTMLUnknownElement');
  },'Basic tests');

  for(let el of elements.children) {
    test(() => {
      assert_true(!!el);
      assert_true(popupVisible(el));
      assert_not_equals(window.getComputedStyle(el).display,'none');
    },`${el.id} should be HTMLUnknownElement and should be stylable/visible`);
  }

  function supportsPopup() {
    return Element.prototype.hasOwnProperty("popup");
  }

  test(() => {
    assert_false(supportsPopup());
  },`The popup IDL attribute should not be present on Element`);

</script>
